+++ custom_components/custom_components_auditor/entity.py
@@ -2,6 +2,7 @@
 
 from __future__ import annotations
 
+from homeassistant.const import MATCH_ALL
 from homeassistant.core import callback
 from homeassistant.helpers.device_registry import DeviceInfo
 from homeassistant.helpers.entity import Entity
@@ -15,6 +16,7 @@ class AuditorEntity(Entity):
 
     _attr_has_entity_name = True
     _attr_should_poll = False
+    _unrecorded_attributes = frozenset({MATCH_ALL})
 
     def __init__(
         self,